+++ src/pages/OAuthPage.tsx
@@ -1,13 +1,13 @@
 // pages/OAuthPage.tsx
 import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
+import instance from "../api/axiosInstance";
 import useUserStore from "../store/UserStore";
 import { fetchUserInfo } from "../api/userInfo";
-import loading from "../assets/loading.gif";
 
 const OAuthPage = () => {
   const navigate = useNavigate();
-  const { setUser } = useUserStore();
+  const { username, ImgUrl, userId, setUser } = useUserStore();
 
   const getCodeFromUrl = (): string | null => {
     const params = new URLSearchParams(window.location.search);
@@ -15,6 +15,30 @@ const OAuthPage = () => {
     console.log("code: ", params.get("code"));
 
     return params.get("code");
+  };
+
+  const handleLogin = async (code: string) => {
+    try {
+      const response = await instance.get(`/api/v1/auth/code/google`, {
+        params: { code },
+        withCredentials: true,
+      });
+
+      console.log("로그인 성공:", response.data);
+
+      localStorage.setItem("accessToken", response.data.data.accessToken);
+      localStorage.setItem("refreshToken", response.data.data.refreshToken);
+      
+      navigate("/");
+    } catch (error: any) {
+      console.error("로그인 실패:", error);
+
+      if (error.response?.status === 401) {
+        navigate("/signup");
+      } else {
+        alert("로그인 처리 중 오류가 발생했습니다.");
+      }
+    }
   };
 
   const getUserData = async () => {
@@ -31,8 +55,7 @@ const OAuthPage = () => {
         email: "",
         ImgUrl: userData.profileImg,
         userId: userData.userId,
-      });      
-      navigate("/");
+      });
     } catch (error) {
       console.error("사용자 정보 가져오기 실패:", error);
     }
@@ -43,14 +66,13 @@ const OAuthPage = () => {
     const code = getCodeFromUrl();
     if (code) {
       console.log("Google Authorization Code:", code);
+      handleLogin(code);
       getUserData();
       console.log()
     }
   }, [setUser]);
 
-  return  <div className="flex items-center justify-center">
-  <img src={loading} />
-</div>;
+  return <div>{/** 200이 오면 바로 로그인 페이지로 */}</div>;
 };
 
 export default OAuthPage;
