+++ src/pages/SignupPage.tsx
@@ -1,0 +1,40 @@
+import { useState } from 'react';
+import shootLogo from '../assets/shoot/shootLogo.png';
+import ConnectFigma from '../components/signup/ConnectFigma';
+import SignupMain from '../components/signup/SignupMain';
+import ConnectDiscord from '../components/signup/ConnectDiscord';
+
+const SignupPage = () => {
+    
+    const [isSignupMainVisible, setIsSignupMainVisible] = useState(true);
+    const [isConnectFigma, setIsConnectFigma] = useState(true);
+
+    const handleSignupButtonClick = () => {
+        setIsSignupMainVisible(false);
+    };
+
+    const handleConnectFigmaButtonClick = () => {
+        setIsConnectFigma(false);
+    };
+
+    return (
+        <div className="flex flex-col items-center justify-center w-full h-screen">
+            <img 
+                src={shootLogo} 
+                alt="shootLogo" 
+                style={{ width: "315px", height: "66.23px" , top: '300px'}} 
+            />
+            {isSignupMainVisible ? (
+                <SignupMain onButtonClick={handleSignupButtonClick} />
+            ) : (
+                isConnectFigma ? (
+                    <ConnectFigma onButtonClick={handleConnectFigmaButtonClick} />
+                ) : (
+                    <ConnectDiscord />
+                )
+            )}
+        </div>
+    );
+};
+
+export default SignupPage;
