+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/env.d.ts","./src/main.tsx","./src/api/axiosinstance.tsx","./src/api/index.ts","./src/api/userinfo.tsx","./src/components/header.tsx","./src/pages/loginpage.tsx","./src/pages/mainpage.tsx","./src/pages/oauthpage.tsx","./src/pages/profilepage.tsx","./src/pages/signup/signuppage.tsx","./src/pages/signup/components/connectdiscord.tsx","./src/pages/signup/components/connectfigma.tsx","./src/pages/signup/components/signupmain.tsx","./src/store/userstore.ts","./src/styles/color.ts","./src/styles/typography.ts"],"version":"5.6.3"}
+{"root":["./src/app.tsx","./src/env.d.ts","./src/main.tsx","./src/api/axiosinstance.tsx","./src/api/userinfo.tsx","./src/components/header.tsx","./src/components/userprofile.tsx","./src/components/signup/connectdiscord.tsx","./src/components/signup/connectfigma.tsx","./src/components/signup/signupmain.tsx","./src/pages/loginpage.tsx","./src/pages/mainpage.tsx","./src/pages/oauthpage.tsx","./src/pages/signuppage.tsx","./src/pages/userpage.tsx","./src/store/userstore.ts","./src/styles/color.ts","./src/styles/typography.ts"],"version":"5.6.3"}
